t/exact_test_sampler: make _generate_in_parallel work nicely with an idle task
Until now, we assume that the main process was actively occupied with
simulating more values for exact testing, and the process pool was
only an optimistic helper.  That won't be the case when we use the
same ExactTestSampler code in a gRPC service.

Make sure to sleep instead of busy polling in ExactTestSampler,
and make sure the pool is never empty.

TESTED=existing tests, t/bench.sh with only the pool computing enabled.